# [corefoundation] Cache `kCFNull` to avoid native calls