@spouliot 2022-05-27
Noticed the `X64 RyuJIT` in the logs :( even if the project was changed to `osx-arm64`. Turns out the pre-test process is **Intel** not **Apple**
Not that I expect the ratio of being very different...