@vs-mobiletools-engineering-service2 2022-05-27
# :computer: [PR Build] Tests on macOS Mac Catalina (10.15) passed :computer:

:white_check_mark: **All** tests on macOS Mac Catalina (10.15) passed.

[Pipeline](https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=6213171&view=ms.vss-test-web.test-result-details) on Agent 
Hash: [5bcfb8ded7a762567e2a824aacf1686a05382a41](https://github.com/xamarin/xamarin-macios/pull/15146/commits/5bcfb8ded7a762567e2a824aacf1686a05382a41)